Add example external CA data provided by Dustin Mitchell
Without this patch applied we don't have an easy way to create a set of
CA and SSL certificates that are representative of how end users are
currently using external CA's with Puppet.

This is a problem because we're left to guess how users are using Puppet
while trying to support external CA's.

This patch adds additional data created by Dustin's `certchain.sh`
script which are not being actively used in the acceptance tests.
They're included in Puppet in an effort to reduce roadblocks to
augmenting the existing acceptance tests.