(maint) Make external CA acceptance test work with SELinux
Without this patch the External CA acceptance tests are failing in CI.
This is a problem because we need to automatically validate that our
documented external CA configuration remains in order.

The root cause of the problem is that SELinux is enabled and enforcing.
This patch addresses the problem by setting the appropriate permissions
on TCP port 8141 and the document root used for testing.