(PUP-672) TrustedInformation handles a nil certificate.
Acceptance caught an issue with an external ca not passing the
certificate info in the request (the
acceptance/tests/external_ca_support/apache_external_root_ca.rb does
this because its SSLOptions does not have +ExportCertData in
acceptance/tests/external_ca_support/fixtures/httpd.conf).

So we are modifying the TrustedInformation class to handle a nil
certificate without failure, but issue an info level warning in the
logs.

This also modifies the acceptance test to ensure a teardown, as failure
to reach the clean up code is breaking following tests when the httpd
server gets left up as master.