(PE-3909) Wrap try/catch for error test around bootstrap
This commit updates the `test-error-during-init` jruby-puppet-service
test to wrap a try/catch for an expected initialization exception around
the bootstrap for the application rather than just the run-app phase.

This change is necessary because an exception may be propagated up during
the bootstrap phase rather than the run-app phase if the call to
`create-jruby-instance` from a future during initialization happens to
be really fast.