(PE-3952) Set the ssl-crl-path from Puppet core to the webserver service
This commit retrieves the `cacrl` path from core Puppet and uses this
value to override the `ssl-crl-path` setting that the webserver service
uses.

This commit bumps up the dependency on the
trapperkeeper-webservices-jetty9 package to 0.5.2, the first version of
this package which supports configuration of a CRL, and the dependency
on the trapperkeeper package to 0.4.2, the version that the new
trapperkeeper-webservices-jetty9 package depends upon.

This commit also bumps the `http-client` and `certificate-authority`
package dependencies up to the latest available at the time of the
commit, 0.1.7 and 0.1.5 respectively.

In conjunction with updating these dependencies, the config.rb class had
to be changed to point to new location of the `pems_to_ssl_context`
helper method that it uses.  This method is now in the
`CertificateAuthority` package.

Additionally, some certificate-related files are included in this
commit for testing.  The `localhost-compromised` certificate represents
a certificate which was issued by the CA but was later revoked by the
CA.  The serial number of the `localhost-compromised` certificate is
represented in the updated `ca_crl.pem` file.