(PE-3590) Call kitchensink cn-for-dn to get CN from cert DN
This commit changes the `get-cert-common-name` function to use the
`cn-for-dn` function in clj-kitchensink to get the common name from the
certificate's distinguished name, as opposed to reimplementing the same
logic locally.

This commit also removes the warning that previously would have been
logged in the event that a client certificate was not provided in a
request routed to JRuby.