(PE-4304) Removed support for marshaling param vectors to Ruby Arrays
This commit removes the support that the prior commit had for marshalling
array-type HTTP parameter values in a Clojure Ring/Compojure request map
into a Ruby Array during request handling.  We don't have a known working
case at present that would appear to require this and are concerned about
the extra processing overhead that this conversion would have otherwise
added to every request.