(PE-3826) Implement naive autosign
This commit modifies the CA service to honor Puppet's 'autosign' setting,
but only for boolean values (whitelist & executable files coming soon).
If the value of autosign is true, the CSR will be signed and the certificate written to disk.
Note this is the current behavior of our CA, which basically assumes autosign is always true.
If autosign is false, the CSR will be written to disk instead.