(PE-4463) Address PR feedback
This commit merges the CaFilePaths and CaSettings schemas.
This allows us to pass the ca-name around in the settings hash
rather than alongside it as a separate argument.

This also removes pre-conditions in favor of schema validation,
which in this case serve only as documentation.

Also explicitly create the certdir when initializing the master.