(PE-3178) Don't return serialized Ruby YAML
This commit changes the HTTP response for the PUT CSR request
to an empty plaintext body rather than the unused serialized Ruby YAML.

After these changes were made, the certificate expiration logic was exposed
as being incorrect, and no longer used right now. I deleted the unused code
and removed the relevant portion of the test and put a TODO comment in place
for when we properly support ca-ttl. I pasted the code in the JIRA ticket
(PE-3173) so we can reuse the logic if necessary when we return to it.