(PUP-417) Make use of QENG-188 to restart puppet in acceptance
Versions of Beaker prior to the patch in QENG-188 relied on manually
starting and stopping puppet with a `puppet master *args` commandline
invocation followed by a `kill $puppet_pid`. To better mirror manual
smoke testing, we want to be able to start and stop puppet using the
service scripts provided by the installation packages, using `puppet
resource service` calls. We also want to be able to test Puppet from
packages that install with Passenger, relying on an apache restart.

These changes are all in Beaker, in QENG-188, but this commit is needed
to adjust puppet acceptance's setup steps to differentiate between a git
installation, a foss package installation, and a foss package
installation with passenger. Several tests were adjusted as well to be
agnostic between foss webrick/passenger installations, wherever
possible.

This commit also pins to Beaker ~> 1.17, which will pick up future
versions of Beaker < 2.0.  Beaker is smoke testing PRs against Puppet
acceptance, among other projects, so I think it's safe to stay
abreast of releases in the 1.y series.