(PE-4751) Use Puppet's spec_helper when running rspec tests
When running puppet's spec tests, there is a lot of logic in
the spec_helper.rb file that does setup and teardown around
all of the tests. It's particularly important for things like
resetting all of the puppet settings between tests, so that
if a test changes the settings, it won't pollute other tests.
In many cases you will get an explicit error about the
settings having already been initialized if you try to run
tests without the spec_helper.

In order to use their spec_helper, we also have to install
all of the gems specified in their Gemfile.

This commit modifies our Rakefile to do horrible things with
bundler to accomplish that. This would also be a prerequisite
for ever trying to run the regular puppet spec tests via our
lein JRuby.