Remove CHANGELOG
Reviewing the CHANGELOG in preparation for an internal 0.1.6 release,
it was clear that we're not quite to the point where the details
that would go into the CHANGELOG would be of any use to users.

This commit deletes it for now.  We will add it back in when we
get closer to being ready for an external release.