(PE-3864) delete extension filtering
Previously, the extensions were filtered from the CSR before signing the
cert, so that any non-Puppet extensions would be omitted.  This is no
longer necessary, since there is now validation that will result in an
error being returned to the client if the CSR contains any disallowed
extensions.

Delete this code and the corresponding test case.