(PE-5437) Get ruby load path from `os-settings`
Because of some issues relating to packaging, we needed to create
a new config file / section called `os-settings`, which we
can populate with the correct ruby load path for the puppet
ruby code at package/install time.

This commit simply changes the code to actually read the setting
from this location, rather than from its previous location
(which was in the `jruby-puppet` config section).