(PE-5567) Return string rather than stream from http requests
The HTTP client we're registering with Puppet was returning an
InputStream for the body, but the rest of the Ruby code is
expecting a String.

In the future it might be nice to make this more flexible, and
consume it as a stream in some cases when possible, but for now
we may as well read it in to a String since that's what is
required by all of the callers.