(PE-5635) Remove `override-webserver-settings` call
This commit removes the call that puppet server previously would make to
the `override-webserver-settings!` function in order to override those
settings with corresponding values from core Ruby Puppet code.  This
change allows the webserver settings in the webserver conf file to be
preserved during run-time execution of the service.  In order for the
same values to be used at the webserver layer and core Ruby Puppet
layer, the settings in the configuration files for each would now have to
be "manually" set to be the same.