(PE-5635) Add SSL settings to dev-resources/puppet-server.conf
This commit adds the SSL cert/key-related settings to the
dev-resources/puppet-server.conf file's webserver section that
are needed in order for the service to be booted.  This is needed
because the settings are no longer automatically inferred from Ruby
Puppet now that the `override-webserver-settings!` call is no longer
being made.