(PE-5084) refactor validate-duplicate-cert-policy!
... so that it now takes the CSR object instead of just the subject.
This will be needed to check the status (revoked/signed) of the CSR so
that we can generate the proper error message.

This commit also adds a new test file - a CSR for a subject named
'localhost'.  This test file is now used in a test case which verifies
that 'proces-csr-submission!' throws a given exception when there
is already a signed certificate in-place for the CSR subject, as well as
another test case for the behavior of the 'allow-duplicate-certs'
setting.

These test cases had to be updated due to the fact that
'validate-duplicate-cert-policy!' now reads the subject from the CSR
instead of from the 'subject' value passed to 'process-csr-submission!'.