(PE-5635) Revert removal of call to `override-webserver-settings`
This commit reverts all of the previous work done to initially remove
the call to `override-webserver-settings` for PE-5635.  The previous
work would have introduced the need for all of the SSL-related settings
to be added to the `webserver.conf` file in packaging and the logistics
of getting that right for the initial release seemed too difficult to
work out.