(PE-5510) Honor the Puppet `ssl-server-ca-auth` setting
This commit causes the webserver setting for `ssl-ca-cert` to be
overridden by the Puppet `ssl-server-ca-auth` setting if it is set in the
puppet.conf file.  If the `ssl-server-ca-auth` setting is not set, the
`ssl-ca-cert` value will be overridden by the Puppet `cacert` setting
instead.  This behavior is intended to match the behavior described for
the `ssl_server_ca_auth` setting at
https://docs.puppetlabs.com/references/latest/configuration.html#sslservercaauth.
The two settings allow the user to distinguish between a set of CA certs
to be used for authenticating clients vs. the CA cert used by the Puppet
Internal CA for signing certificates.

This commit also adds an exclusion of `joda-time` from `clj-time`.
Leiningen found that `clj-time` depended upon an older version of
`joda-time`, 2.2, than `jruby-core` depends upon, 2.3.