(PE-5703) Remove override call when webserver settings predefined
This commit removes the call that the Puppet Server Config service would
make to `override-webserver-settings!` when the SSL cert settings were
predefined in the Trapperkeeper webserver conf file.  This change allows
the TK SSL cert settings to be omitted and inferred from puppet.conf
settings by default but for the TK SSL cert settings to trump the
puppet.conf settings for special cases, e.g., when the user wants to
configure the master to use SSL-related files from an external
(non-Puppet) CA.

This commit adjusts the log messages in order to indicate whether the
override call is being made or not while still allowing the values of
all of the keys read from core Puppet code to be logged in debug mode.