(PE-3601) Get rid of `scratch` directory
This commit provides a major cleanup of how we have been organizing
files that we are using for dev and test purposes.

Prior to this commit, we had a big glob of files in `dev-resources`,
and it was difficult or impossible to reason about which ones were
being used by which tests (or indeed whether they were being used
at all).  There were also a bunch of files that weren't really
for tests, per se, but rather for use when doing development in the
repl, etc.

On top of that, we had some shell scripts that would create
a directory called `scratch` and copy a bunch of files there,
and there were certain tests and development patterns that
would not work properly without these files.  This was also
very hard to reason about.

Lastly, we had a directory inside of `dev-resources` in our
`.gitignore`, because certain tests were writing files to it.

This commit attempts to remediate all of the above.  The major
themes toward that goal:

* Get rid of the bin scripts and the scratch directory altogether,
  and make sure that nothing relies on them.
* Create a new `dev` directory for all developer-related files,
  such as REPL code, developer logging config, sample config files,
  etc.
* Reorganize `dev-resources` so that it conforms to the following
  rules:
     * All files contained therein are fixture/data files used by
       tests
     * Files for any given test must reside in a directory that
       matches the namespace of the test, so that it is immediately
       clear where to look if you need to know what a file is there
       for.
* Refactor all necessary tests to comply with above changes

As a result of these changes, there were several files deleted from
`dev-resources` because they did not appear to actually be used by
any existing tests.