(PUP-3165) Ensure agent run on master during pre_suite.
This will ensure that `/var/lib/puppet/reports/<master-fqdn>` gets created by
the master process running as the master user. This is necessary because
otherwise some early testcase may get a chance to run 'puppet apply' which will
create the same directory as root, thereby making it unavailable to the master
user.

Signed-off-by: Wayne <wayne@puppetlabs.com>