(PE-5704) /certificate_status - validate CSR
CSRs are now validated before signed via /certificate_status.  If the
validation fails, a 409 is returned, and the error message returned to
the client is the same error message that would have been returned
when the CSR was originally received by the master.

This commit also tidies-up the :conflict? and :handle-conflict parts of
the liberator resource by storing the appropriate error message in the
context when the conflict is detected.