(PE-4308) Awesome fixes to the CA ring handler
Remove the 'test-pson-as-json' and 'wrap-json-body' middlewares.  They
were added for /certificate_status but broke other things, because we
have to mount both the CA and the master's ring handlers at /, because
everything is awesome!

This commit also changes /certificate_status to ignore the
'Content-type' header on incoming requests.  There was no real reaon to
honor it in the first place - we can either parse the request as JSON
and use it, or we can't do anything with the request.

As a result of this, the ring-json dependency is removed.