(PE-3601) Fix tests that rely on vardir
Prior to this commit, there were a few tests that were using the
puppet `vardir` setting implicitly, which evaluates to ~/.puppet
if you do not set it explicitly.

This commit adds a test fixture that can be used to deploy a
`puppet.conf` file for a given test, and deletes it afterwards.
It also adds code to call this fixture for the handful of tests
that require it.