(PE-5888) Default Content-Type in response to text/pson if not set
This commit forces the Content-Type in the certificate_status*
responses to be "text/pson" if the Content-Type is not already set at the
point the representation is converted into a Ring response.  This is
needed because Liberator by default would have added an empty
Content-Type if the Accept header in the request were "pson".