(PE-6073) Fix rake spec tests
This commit changes the expected version in the puppet version spec test
to 3.7.1 so that the test is sync with current puppet submodule being
selected by the project.

This commit also wraps the `RSpec::Core::Runner.run` call that
run_specs.rb makes in an exit call so that the return code from the
`run` call is propagated out.  Previously, the result from the `run`
call was swallowed, causing the calling application to always have an
exit code of 0 even if tests had failed.