(RE-3046) Remove usermod, chmod from preinst
Previously the preinst for puppet-server modified the vardir of puppet
to be world traversable and world readable, which is a problem because
of the private data that is often stored in the vardir. This was
actually unnecessary because puppet-server runs as the puppet user. This
code was a leftover from when it was running as the puppetserver user
that was added to the puppet group. This commit removes the usermod that
is no longer necessary and removes the chmods which are also not
necessary.