(SERVER-86) Don't use CA public key file for autosign/revoke
This commit removes the usages of $capub from the autosign and revoke
functions. They will now extract the public key from the $cacert file,
which we can assume will always be available to those functions, unlike
the $capub file.

The only time that we should be reading/writing the $capub file is
during CA initialization.