(SERVER-112) Clean up puppet server ruby config
Prior to this commit, we were passing in the optional profiler
object from the Clojure code to the Ruby code as a first-class
argument to the constructor of our Ruby `Master` class.

It's becoming clear that we're going to have more of these
Puppet-Server-specific config options that we have to pass through
over time (including some things that will be coming soon in order
to complete SERVER-112), so this commit refactors the code to
pass in a map containing *all* Puppet Server settings as a single
argument.

It also makes an initial effort to start moving the handling of
some of those settings into the Puppet::Server::Config class,
and tries to group them together a bit more inside the body
of the `Master` constructor.