(SERVER-112) Introduce bootstrap testutils, improve bootstrap test
Prior to this commit, the bootstrap test wasn't actually doing
much work.  It would create a trapperkeeper app to make sure that
there were no failures in just loading the service *namespaces*,
but it did not call any of the lifecycle functions of any of the
services, so would not actually do much in terms of validating
that the app could be run.

This commit improves the test to make it actually do a full startup
of the app based on the sample config files.

It also introduces a bootstrap_testutils namespace, which contains
some test utility code relating to booting up a puppet server
instance during testing.  This will be useful for some forthcoming
tests that I am writing related to testing puppet's directory
environments.