(SERVER-151) Move environment fns from jruby-testutils to service
We originally defined some functions for interacting with the
environment cache state in jruby-testutils.  This commit moves
the logic into the `JRubyPuppetService` instead; this will be necessary
in order to wire up the HTTP endpoint, since we will only have
access to the service and not to the entire TK app.