(MAINT) Make user repl config dynamic
This commit modifies `user_repl.clj` and the sample `user.clj`
file to use a function instead of a var for the user-provided
config.

Prior to this, the config was read from disk as a var when the
REPL was first launched, and never re-loaded after that.  This
meant that doing a `reset` would not pick up changes you might
have made to the config files on disk; you needed to restart
the REPL to pick up those changes.

Now, the user config is acquired via a function call, which
means that your config files will be reloaded each time you
call `reset`.