(SERVER-85) Sync CA CRL to Host CRL at startup
This commit changes the master to, at each startup, copy the file
pointed at by the Puppet `cacrl` setting, if it exists, over to the
file pointed at by the Puppet `hostcrl` setting.

This commit also changes the `override-webserver-setting!` logic to
use the file at the `hostcrl` setting rather than the one at the `cacrl`
setting to determine which file the webserver uses to validate
certificates for connecting clients.  This was done based on the fact
that a "standalone" master -- one hosted without a corresponding active
CA in the same container -- would have used the `hostcrl` setting under
the Ruby master.  With the logic in place to copy the `cacrl` to the
`hostcrl` when a real CA is in place, the latest CRL, when the master
and CA are running in the same container, should now be in place at
service startup.