(maint) Add in support for the packaging repo
Since puppet-server is built with ezbake, this addition is not
necessary. However, we do want to be able to create release tickets for
the puppet-server project when the time comes for a release. It does
seem a little silly to add in support for the entire packaging repo just
to create release tickets, but ultimately that funcionality will be good
to have easy access to.

Note, the yaml files included in this pull request are the minimum
viable to generat the release tickets. A lot of the other tasks and
automation available through the packaging repo will not work because of
what is left out. Again, if those tasks are needed, they should be run
with EZBake.

In order to create tickets, in the puppet server project, run the
following:

`rake package:implode package:bootstrap`
`rake pl:tickets BUILDER=melissa DEVELOPER=$you WRITER=jean
RELEASE=1.0.0 DATE=2014-12-15 JIRA_USER=$you PROJECT=SERVER`

The DEVELOPER corresponds to the persion who is going to coordinate the
release. The BUILDER is the Release Engineer working on the release, and
the JIRA_USER is the person who is creating the tickets. It is important
that the names all correspond to actual jira usernames.

More documentation is available at
https://github.com/puppetlabs/packaging/blob/master/tasks/tickets.rake#L408