(SERVER-221) Initialize run_mode earlier
During some recent refactorings, we accidentally changed the behavior
for how we initialize Puppet's run_mode.  We weren't initializing
it as early as we had done in the past, so it was causing a very
minor and hard-to-repro bug in certain edge-case situations (such
as when a user had a puppet setting defined in both the `[main]`
and `[master]` sections of their puppet.conf file.

This commit moves the initialization up to happen a little earlier,
so that we'll be in `master` run_mode from the get-go.