(SERVER-151) Address PR feedback
Add comment to jruby-pool-resource similar to that for the
environment-cache-resource. Change resources test directory for
jruby-pool-int-test. Remove unnecessary Accept header from
test of the jruby pool endpoint. Add tests to ensure that the
jruby-pool endpoint tolerates an 'Accept: */*' header and that
it denies access when a cert is not on the whitelist.