(SERVER-245) Avoid Puppet 4 being used as a dependency
Without this patch puppet-server allows any version of Puppet >= 3.7.3.
This is a problem because Puppet 4 is on the horizon and will introduce
changes in behavior incompatible with behaviors we depend on.

This patch addresses the problem by specifying that RedHat and Debian style
packages should only use puppet packages >= 3.7.3 and < 4.0.0 to satisfy
the dependencies for puppet-server.