(SERVER-250) Address PR feedback
Explain what "flushing" the JRuby pool means. Add a Use Case.
Explain that this is an experimental feature and that the
performance impact is unknown. Add disclaimer about the Master
being unable to handle requests until the first interpreter
is initialized. Explain why you'd want to flush the pool instead
of the environment cache. Prepare document for the website.