(docs) Make curl examples use certificates
Since we're expecting a bunch of people to kick the environment cache directly
from the CLI, it makes sense to have the curl examples be as close to
copy-pasteable as we can get. Which isn't very close, since you need to sub in
paths to certificates and edit your config files, but this at least exposes the
correct options so people can skip reading the curl man page.