Implements a TK service to access fileserver.conf mounts
This service parses puppet fileserver.conf and keeps the list
of active mounts.

It also exports functions to find a mount matching a given path, and
to check a given mount access is allowed for a request.

It doesn't yet support the special "plugins" and "modules" mounts, nor
acl mount checking.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>