(MAINT) Normalize subcommand environment handling
Without this patch each of the `gem`, `ruby`, and `irb` subcommands
handle environment variables in different ways.  This is a problem
because the project as a whole strives to manage environment variables
such as GEM_HOME and GEM_PATH as robustly as possible.  It is also
problematic to manage additional environment variables such as
JARS_NO_REQUIRE because the variable must be added to each subcommand's
implementation.

This patch addresses the problem by moving the gem/cli-gem-environment
function into subcommand/environment.  The function signature remains
the same.  The behavior changes slightly to explicitly manage all of the
environment variables puppetserver cares about from a robustness
perspective.  These are GEM_HOME, GEM_PATH, RUBYLIB, RUBYOPT, and
RUBY_OPTS.

This change also provides a single location to manage additional
variables such as JARS_NO_REQUIRE for use with all subcommands.