(SERVER-247) Add jira-ruby to Gemfile
Without this patch the `pl:tickets` rake task fails because the
jira-ruby dependency is missing.  This patch addresses the problem by
adding jira-ruby to the project Gemfile.  The dependency will be
available following `bundle install`.

The library is in the development group to allow the dependency to be
filtered using `bundle install --without development`.

This patch also moves rake out of the test group because the rake
dependency is always required when running rake tasks for development
and testing.