(SERVER-252) Add acceptance tests for irb and ruby subcommands
Without this patch there are no acceptance level tests for the ruby and
irb subcommands as they relate to the `gem` subcommand.  This is a
problem because we need to ensure these commands behave as expected,
particularly when the `gem` subcommand is used to install third party
libraries.

This patch addresses the problem by installing the `hiera-eyaml` gem and
then checking that the `ruby` and `irb` subcommand can `require 'eyaml'`.

This patch also adds basic testing of the behavior of GEM_HOME and
FOO_DEBUG with regard to the irb and ruby subcommands in addition to
ensuring `require 'puppet'` works in both subcommands.