(PUP-3526) Make default CA prefix '/puppet-ca'
Make the default CA url prefix '/puppet-ca', rather than '/ca'. Since the
master url prefix is '/puppet', and since the routing will match '/puppet-ca'
to '/puppet', this also means tweaking how we register routes, so that the
prefix is registered with a slash on the end and the version strings do not
begin with slashes.