(SERVER-272) Add version to CA and remove environment reference
The CA routes were split out completely from the master routes (see PUP-3526).
They are now at a different prefix and version than the master routes in
Puppet's code. This means, that in Puppet Server they can have their own
prefix (`/puppet-ca`) - they no longer need to be mounted at `/`.

In addition, the environment has been moved from the first part of the path to
a query parameter. The environment is never used in the CA routes, so this
query parameter is now ignored completely.