(SERVER-252) Fix bad irb and ruby subcommand tests
Without this patch the tests are failing because of bad logic in the
tests themselves.  This patch addresses the problem by correcting the
logic in the tests, specifically /BAD/ was matching in the irb tests
because irb echos input to standard output.  /^BAD$/ will not match the
echoed input.  The ruby test was failing because string interpolation
was happening in the beaker process when it was intended to be passed
through beaker to the ruby subcommand being tested.  This patch
addresses the problem by avoiding string interpolation entirely and
using concatenation instead.