(SERVER-272) Add not found handler for master and ca routes
Previously both the master and CA routes had to be mounted at `/` because of
Puppet's url structure. Since the environment was always the first part of the
path, and since environment could be almost anything, this meant that we could
not really detect when there was no route to match a request and issue a `Not
Found` error. Since these routes now have prefixes and since the environment
is a query parameter, we can now detect this and issue a 404.