(SERVER-325) Switch from ArrayBlockingQueue to LinkedBlockingDeque
Prior to this commit, we were using a FIFO strategy for allocating
JRuby instances to handle requests.  This has the side effect that, at
any given time, all of the JRuby instances in the pool will have
handled roughly the same number of requests.

This is problematic if we want to implement a
`max-requests-per-jruby-instance` setting, because it will mean that
we need to flush all of the JRuby instances at roughly the same time,
which will be a big performance hit and will not leave as many instances
in the pool to handle other requests while we're flushing.

To remediate this, we need to switch to a LIFO model.  This will result
in the first few JRuby instances in the stack handling a larger number of
requests than the remaining instances, so they will hit their max request
count sooner, and can be flushed invidually without flushing the entire
pool.

To that end, this commit simply changes the data type that we use to
manage the JRuby pool, from `ArrayBlockingQueue` (which only supports FIFO
semantics), to `LinkedBlockingDeque`, which supports both FIFO and LIFO.