(SERVER-325) Switch from FIFO to LIFO for allocating JRubies
As discussed in the previous commit, we need to move from
FIFO to LIFO for our JRuby allocation in order to be able
to flush single instances efficiently when they've handled
a certain number of requests.  This commit modifies the
borrow/return functions to use a LIFO strategy, and alters
some tests accordingly.