Update options for centos 6 jenkins test
Specifically for various paths. Not entirely sure these are needed, they were
mostly temporary hacks to get the tests running. Either the defaults for these
will need to be changed in beaker (which may have already been done, but not
yet released), or we'll need to add these to all of the configs for jenkins.