(MAINT) Fix broken config service test
There was a test in the PuppetServerConfigService that was relying
on the string representation of a set to be printed in a certain
order, even though the data structure makes no guarantees about
sort order.  It seems that this test started failing when we
updated to Clojure 1.6.0 (perhaps intermittently), probably due
to changes in the underlying implementation of the set data structure.

This commit adds an explicit sort to the set before representing
it as a string, so that the resulting string is deterministic.  It
also updates the test to account for this.

The test was also not specifying a confdir, so it was attempting
to access the default user vardir (`~/.puppet`).  This commit addresses
that problem as well.

There was also an issue with the project dependencies bringing in
conflicting versions of `commons-codec` in certain cases, so this
commit pins us to an explicit version of `commons-codec`.

Conflicts:
	test/unit/puppetlabs/services/config/puppet_server_config_service_test.clj